+++ examples/web-automation/tasks/__init__.py
@@ -1,0 +1,9 @@
+"""Web automation tasks package."""
+
+from .extract_title import ExtractPageTitleTask
+from .navigate import NavigateTask
+
+__all__ = [
+    "NavigateTask",
+    "ExtractPageTitleTask",
+]